Add database config classes
These classes will be used to fetch information about the database that
the Discraft bot will store data in. It was made its own class (and
configuration file) to help keep things simple in both code and config
file structure. This may be undone in the future if it proves to
be trivial to merge DBConfig and DiscraftBotConfig into one class.