Add multi-language support for DiscraftLang
This took me way to long to write, but it should work decently well.
This is definitely one of those things I'm going to modularize in the
future into STD I/O commons, like the config file stuff that I earlier.
I'll make changes to this as needed, but now it should be possible to
fetch localed messages for users based on their specified language.