Tweak how languages are loaded in DiscraftLang
I thought the previous way was a bit ugly and required some redundant
information to be placed into the JSON file. The "langs" key is now an
array, and contains JSON objects. These objects now contain the ID,
which serve as the basis for the name of the lang folder.