Allow Discord commands to throw exceptions
I made this change since I thought it would be better to just let
exceptions outside of the developers direct control (IO exceptions, SQL
exceptions, etc.) just get caught by the command executor and then
handle it in a uniform way for all commands. That being, notifying the
user an issue occured and storing the error log somewhere for later.