Snap back to reality
Okay so I know this is an atrocious commit but I have no idea what
happened since I last touched the source code for this. I came back, saw
everything in ruins, git status showing GARBFAHFASDFLKJAMWIUJH when I
ran it, and there was a bunch of compile time errors. I fixed this, and
got lazy when doing this commit.